// src/app/tabs/shared/models/credit-note-transaction.model.ts
export class CreditNoteTransaction {
    rowNumber?: number;
    id?: number;
    accountHeadName?: string;
    fixedAmount?: number;
    fixedAmountLocal?: string;
    formulaFieldName?: string;
    operatorName?: string;
    formulaDescription?: string;
    operatorDescription?: string;
    creditNoteMasterId?: number;
    billHeadId?: number;
    taxId?: number;
    headDisplay?: string;
    creditNoteAmount?: number;
    creditNoteAmountLocal?: string;
    transactionMode?: string;
    isVAT?: boolean;
    disabled?: boolean;
    position?: number;
    value?: number;
}